+++ jdaviz/configs/default/plugins/gaussian_smooth/gaussian_smooth.py
@@ -71,6 +71,8 @@ class GaussianSmooth(PluginTemplateMixin, DatasetSelectMixin, AddResultsMixin):
             self.dataset._viewers = [self._default_spectrum_viewer_reference_name]
             self.dataset._clear_cache()
 
+        self.dataset.add_filter('not_from_this_plugin')
+
         self.mode = SelectPluginComponent(self,
                                           items='mode_items',
                                           selected='mode_selected',